Move robust_kernels to separate file
