+++ cpp/pybind/pipelines/registration/registration.cpp
@@ -80,25 +80,6 @@ public:
     }
 };
 
-template <class RobustKernelBase = RobustKernel>
-class PyRobustKernelT : public RobustKernelBase {
-public:
-    using RobustKernelBase::RobustKernelBase;
-    RobustKernelType GetRobustKernelType() const override {
-        PYBIND11_OVERLOAD_PURE(RobustKernelType, RobustKernelBase, void);
-    }
-    double Weight(double residual) const override {
-        PYBIND11_OVERLOAD_PURE(double, RobustKernelBase, residual);
-    }
-};
-
-// Type aliases to improve readability
-using PyRobustKernel = PyRobustKernelT<RobustKernel>;
-using PyL2Loss = PyRobustKernelT<L2Loss>;
-using PyL1Loss = PyRobustKernelT<L1Loss>;
-using PyHuberLoss = PyRobustKernelT<HuberLoss>;
-using PyTukeyLoss = PyRobustKernelT<TukeyLoss>;
-
 void pybind_registration_classes(py::module &m) {
     // open3d.registration.ICPConvergenceCriteria
     py::class_<ICPConvergenceCriteria> convergence_criteria(
@@ -489,50 +470,6 @@ must hold true for all edges.)");
                         rr.correspondence_set_.size());
             });
 
-    // open3d.registration.RobustKernel
-    py::class_<RobustKernel, std::shared_ptr<RobustKernel>, PyRobustKernel> rk(
-            m, "RobustKernel",
-            "Base class that models statistical robust kernels. The virtual "
-            "function Weight() must be implemented in subclasses.");
-    rk.def("weight", &RobustKernel::Weight, "residual"_a,
-           "Obtain the correspondent weight for the given residual.");
-    docstring::ClassMethodDocInject(
-            m, "RobustKernel", "weight",
-            {{"residual", "residual from the optimization problem"}});
-
-    // open3d.registration.L2Loss:RobustKernel
-    py::class_<L2Loss, std::shared_ptr<L2Loss>, PyL2Loss, RobustKernel> l2_loss(
-            m, "L2Loss", "L2Loss class");
-    py::detail::bind_copy_functions<L2Loss>(l2_loss);
-
-    // open3d.registration.L1Loss:RobustKernel
-    py::class_<L1Loss, std::shared_ptr<L1Loss>, PyL1Loss, RobustKernel> l1_loss(
-            m, "L1Loss", "L1Loss class");
-    py::detail::bind_copy_functions<L1Loss>(l1_loss);
-
-    // open3d.registration.HuberLoss:RobustKernel
-    py::class_<HuberLoss, std::shared_ptr<HuberLoss>, PyHuberLoss, RobustKernel>
-            h_loss(m, "HuberLoss", "HuberLoss class");
-    py::detail::bind_copy_functions<HuberLoss>(h_loss);
-    h_loss.def(py::init([](double k) { return new HuberLoss(k); }), "k"_a)
-            .def("__repr__",
-                 [](const HuberLoss &rk) {
-                     return std::string("RobustKernel::HuberLoss with k=") +
-                            std::to_string(rk.k_);
-                 })
-            .def_readwrite("k", &HuberLoss::k_, R"(Insert here docstring)");
-
-    // open3d.registration.TukeyLoss:RobustKernel
-    py::class_<TukeyLoss, std::shared_ptr<TukeyLoss>, PyTukeyLoss, RobustKernel>
-            t_loss(m, "TukeyLoss", "TukeyLoss class");
-    py::detail::bind_copy_functions<TukeyLoss>(t_loss);
-    t_loss.def(py::init([](double k) { return new TukeyLoss(k); }), "k"_a)
-            .def("__repr__",
-                 [](const TukeyLoss &rk) {
-                     return std::string("RobustKernel::TukeyLoss with k=") +
-                            std::to_string(rk.k_);
-                 })
-            .def_readwrite("k", &TukeyLoss::k_, R"(Insert here docstring)");
 }
 
 // Registration functions have similar arguments, sharing arg docstrings
@@ -650,6 +587,7 @@ void pybind_registration(py::module &m) {
     pybind_feature_methods(m_submodule);
     pybind_global_optimization(m_submodule);
     pybind_global_optimization_methods(m_submodule);
+    pybind_robust_kernels(m_submodule);
 }
 
 }  // namespace registration
